alot of changes

+++ SparkyCore/tf2/Spxarky/Spxarky/src/Features/Backtrack/Backtrack.cpp
@@ -294,7 +294,7 @@ void CBacktrack::CleanRecords()
 
 		//const int iOldSize = pRecords.size();
 
-		const int flDeadtime = I::GlobalVars->curtime + GetReal() - m_flMaxUnlag; // int ???
+		const float flDeadtime = I::GlobalVars->curtime + GetReal() - m_flMaxUnlag;
 		if (vRecords.size() > 1 && vRecords.back().m_flSimTime == std::numeric_limits<float>::max())
 			vRecords.pop_back();
 		while (!vRecords.empty())
@@ -328,12 +328,10 @@ void CBacktrack::Store()
 
 void CBacktrack::ResolverUpdate(CBaseEntity* pEntity)
 {
-	/*
 	if (!m_mRecords.contains(pEntity))
 		return;
 
 	m_mRecords[pEntity].clear();
-	*/
 }
 
 void CBacktrack::ReportShot(int iIndex)
